// spfx/role-portal/src/webparts/rolePortal/components/RolePortal.module.scss.d.ts
declare const styles: {
  rolePortal: string;
  header: string;
  kicker: string;
  title: string;
  subtitle: string;
  layout: string;
  nav: string;
  navTitle: string;
  navList: string;
  navLink: string;
  sections: string;
  section: string;
  sectionHeader: string;
  sectionBody: string;
  sectionText: string;
  sectionSubTitle: string;
  externalLinks: string;
  externalIcon: string;
  compact: string;
};

export default styles;